// src/app/models/UsuarioModel.ts
export class UsuarioModel {
    id: number;
    nome: string;
    email: string;
    unidadeId: number;
    unidadePacoteId: number;
    cursoId: number;
    identidade: string;
    celular: string;
    sexo: string;
    img: string;
    ra: string;
    dataNascimento: string;
    codigo: string;
}